third_party/WebKit/LayoutTests/fast/canvas-api/canvas-invalid-strokestyle.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<canvas id="canvas" width="200" height="200"></canvas>
<script>
test(function(t) {
    var ctx = document.getElementById('canvas').getContext('2d');
    ctx.strokeStyle = 'rgb(0, 255, 0)';
    ctx.strokeStyle = 'nonsense';
    ctx.lineWidth = 200;
    ctx.moveTo(0,100);
    ctx.lineTo(200,100);
    ctx.stroke();
    var imageData = ctx.getImageData(0, 0, 200, 200);
    var imgdata = imageData.data;
    assert_equals(imgdata[4], 0);
    assert_equals(imgdata[5], 255);
    assert_equals(imgdata[6], 0);
}, 'Verify correct behavior of canvas on an invalid strokeStyle()');
</script>
</body>
